requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
crontab==1.0.1
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
lxml==4.9.2
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
pymongo==4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.29.0
rsa==4.9
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing==3.7.4.3
typing_extensions==4.6.1
urllib3==1.26.16
uvicorn==0.21.1 
webdriver-manager==3.8.6
wsproto==1.2.0
python-multipart
